requirements.txt
asgiref==3.2.3
Django==3.0.6
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0